refactor(Sidebar): Move type definitions back into components
