+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/SidebarNav.tsx
@@ -1,12 +1,32 @@
 import React from 'react'
-import {
-  SidebarNavItem,
-  SidebarNavItemProps,
-  SidebarNavProps,
-  useSidebar,
-} from './index'
+import { useSidebar } from './index'
 import { warnIfOverwriting } from '../../../helpers'
 import { StyledNav } from './partials'
+import { Nav, NavItem } from '../../../common/Nav'
+import { SidebarNavItem, SidebarNavItemProps } from './SidebarNavItem'
+
+export interface SidebarNavProps extends Nav<NavItem> {
+  /**
+   * Optional handler invoked when `onBlur` event is emitted.
+   */
+  onBlur?: (e: React.FocusEvent<HTMLElement>) => void
+  /**
+   * Optional handler invoked when `onFocus` event is emitted.
+   */
+  onFocus?: (e: React.FocusEvent<HTMLElement>) => void
+  /**
+   * Optional handler invoked when an item is clicked.
+   */
+  onItemClick?: (e: React.MouseEvent<HTMLElement>) => void
+  /**
+   * Optional handler invoked when the `onMouseOut` event is emitted.
+   */
+  onMouseOut?: (e: React.MouseEvent<HTMLElement>) => void
+  /**
+   * Optional handler invoked when the `onMouseOver` event is emitted.
+   */
+  onMouseOver?: (e: React.MouseEvent<HTMLElement>) => void
+}
 
 export function mapNavItem(
   navItem: React.ReactElement<SidebarNavItemProps>,
