+++ packages/react-component-library/src/components/TopLevelNavigation/Sidebar/SidebarNotifications.tsx
@@ -10,7 +10,23 @@ import {
   StyledNotificationsSheet,
   StyledNotificationsSheetButton,
 } from './partials'
-import { SidebarNotificationsProps } from './types'
+import { ComponentWithClass } from '../../../common/ComponentWithClass'
+import { NotificationsProps } from '../NotificationPanel'
+
+export interface SidebarNotificationsProps extends ComponentWithClass {
+  /**
+   * Toggle whether there are unread notifications.
+   */
+  hasUnreadNotification?: boolean
+  /**
+   * Whether the sheet is initially open.
+   */
+  initialIsOpen?: boolean
+  /**
+   * Collection of Notification item components.
+   */
+  notifications?: React.ReactElement<NotificationsProps>
+}
 
 export const SidebarNotifications = ({
   initialIsOpen,
