--- packages/react-component-library/src/components/TopLevelNavigation/Sidebar/types/SidebarUserItemProps.ts
@@ -1,8 +1,0 @@
-import { NavItem } from '../../../../common/Nav'
-
-export interface SidebarUserItemProps extends NavItem {
-  /**
-   * Optional icon to display to the left of the user navigation item.
-   */
-  icon?: React.ReactNode
-}
